// src/app/shared/entities/user-permission.enum.ts
export enum UserPermission {
  SUPER_POWER = 'SUPER_POWER',
  HAS_REPORTS = 'HAS_REPORTS',
  HAS_DASHBOARDS = 'HAS_DASHBOARDS',
  HAS_HOSPITAL_PAGE = 'HAS_HOSPITAL_PAGE',
  HAS_CAREGIVER_PAGE = 'HAS_CAREGIVER_PAGE',
  HAS_BRANCH_PAGE = 'HAS_BRANCH_PAGE',
  HAS_PROVIDER_PAGE = 'HAS_PROVIDER_PAGE',
  HAS_ENROLLMENT_PAGE = 'HAS_ENROLLMENT_PAGE',
  HAS_DEVICE_PAGE = 'HAS_DEVICE_PAGE',
  HAS_CLAIMS_PAGE = 'HAS_CLAIMS_PAGE',
  HAS_TICKETS_PAGE = 'HAS_TICKETS_PAGE',
  HAS_SETTINGS_MODULE = 'HAS_SETTINGS_MODULE',
  HAS_DISEASES_PAGE = 'HAS_DISEASES_PAGE',
  HAS_CPT_PAGE = 'HAS_CPT_PAGE',
  HAS_DEVICE_TYPE_PAGE = 'HAS_DEVICE_TYPE_PAGE',
  CAN_COMMENT_PATIENT = 'CAN_COMMENT_PATIENT',
  CAN_ADD_PATIENT = 'CAN_ADD_PATIENT',
  CAN_CHANGE_ENROLLMENT_TITLE = 'CAN_CHANGE_ENROLLMENT_TITLE',
  HAS_DOCTOR_DASHBOARD = 'HAS_DOCTOR_DASHBOARD',
  CAN_ADD_DEVICE_TO_PATIENT = 'CAN_ADD_DEVICE_TO_PATIENT',
  HAS_PATIENT_REPORT_PAGE = 'HAS_PATIENT_REPORT_PAGE',
  CAN_ADD_PROVIDER = 'CAN_ADD_PROVIDER',
  CAN_CHANGE_PATIENT_REPORTS_TITLE = 'CAN_CHANGE_PATIENT_REPORTS_TITLE',
  HAS_CLINIC_DASHBOARD = 'HAS_CLINIC_DASHBOARD',
  HAS_PRACTICE_DASHBOARD = 'HAS_PRACTICE_DASHBOARD',
  HAS_RPM_ADMIN_DASHBOARD = 'HAS_RPM_ADMIN_DASHBOARD',
  HAS_NOTIFICATION_COUNT = 'HAS_NOTIFICATION_COUNT',
  HAS_NOTIFICATION_PAGE = 'HAS_NOTIFICATION_PAGE',
  HAS_CAREGIVER_DASHBOARD = 'HAS_CAREGIVER_DASHBOARD',
}